Move stuff to folders.
